RagAssignment/requirements.txt
PyPDF2
langchain-text-splitters
sentence-transformers
qdrant-client
google-generativeai
dotenv
streamlit